Modify ETFS layer to create files during write
This is a hack to get around an ETFS limitation that doesn't let you create files. This is so LevelDB can create files in the dbbench/ directory that it creates for testing. Realistically, we let LevelDB directly call into the rumpuser_* layer, so it only goes through this ETFS layer for the metadata files (MANIFEST, etc).